--- voting_election.pro
@@ -1,20 +1,0 @@
-#-------------------------------------------------
-#
-# Project created by QtCreator 2017-01-17T14:34:37
-#
-#-------------------------------------------------
-
-QT       += core gui
-
-greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
-
-TARGET = voting_election
-TEMPLATE = app
-
-
-SOURCES += main.cpp\
-        widget.cpp
-
-HEADERS  += widget.h
-
-FORMS    += widget.ui
